+++ eslint.config.js
@@ -1,4 +1,7 @@
 import antfu from '@antfu/eslint-config'
+import { FlatCompat } from '@eslint/eslintrc'
+
+const compat = new FlatCompat()
 
 export default antfu({
   ignores: ['public/*'],
@@ -8,4 +11,6 @@ export default antfu({
     'react/prop-types': 'off',
     'react/display-name': 'off',
   },
-})
+}, ...compat.config({
+  extends: ['plugin:@tanstack/eslint-plugin-query/recommended'],
+}))
