+++ package.json
@@ -19,17 +19,18 @@
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-tabs": "^1.0.4",
     "@radix-ui/react-toggle": "^1.0.3",
+    "@tanstack/react-query": "^5.22.2",
     "@unocss/reset": "^0.58.5",
     "clsx": "^2.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "recharts": "^2.12.1",
-    "swr": "^2.2.5"
+    "recharts": "^2.12.1"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^2.6.4",
-    "@eslint/eslintrc": "^3.0.1",
+    "@eslint/eslintrc": "^3.0.2",
     "@kuma-ui/vite": "^1.2.2",
+    "@tanstack/eslint-plugin-query": "^5.20.1",
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^14.2.1",
     "@testing-library/react-hooks": "^8.0.1",
